refactor: 헤더 및홈페이지 소개  컨텐츠 내용 리펙토링
